replace custom hook functions with react query queries and mutations; update tests
