+++ src/components/AddTaskContainer/AddTaskContainer.test.tsx
@@ -1,25 +1,38 @@
-import { describe, it, vi } from 'vitest';
+import { describe, it } from 'vitest';
 import AddTaskContainer from './AddTaskContainer';
-import { fireEvent, render, screen, waitFor } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ReactNode } from 'react';
 
 describe('AddTaskContainer', () => {
-  it('calls addNew() with task description as argument when form is submitted', async () => {
-    const addNew = vi.fn();
-    render(<AddTaskContainer addNew={addNew} />);
+  const createWrapper = () => {
+    // creates a new QueryClient for each test
+    const queryClient = new QueryClient({
+      defaultOptions: { queries: { retry: false } },
+    });
+    return function QueryClientProviderWrapper({
+      children,
+    }: {
+      children: ReactNode;
+    }) {
+      return (
+        <QueryClientProvider client={queryClient}>
+          {children}
+        </QueryClientProvider>
+      );
+    };
+  };
 
-    //Type a task description in the input field
+  it('to have been rendered correctly', () => {
+    render(<AddTaskContainer />, { wrapper: createWrapper() });
+
     const inputElement = screen.getByPlaceholderText('New task..');
-    fireEvent.change(inputElement, {
-      target: { value: 'New task description' },
-    });
 
-    //Click the submit button
     const submitButton = screen.getByText('Add');
-    fireEvent.click(submitButton);
 
-    //Assert that addNew() was called with the correct description
-    await waitFor(() => {
-      expect(addNew).toHaveBeenCalledWith('New task description');
-    });
+    //Assert that elements are in the document
+
+    expect(inputElement).toBeInTheDocument();
+    expect(submitButton).toBeInTheDocument();
   });
 });
